[CIv2:POC] Add docker compose file
This docker-compose.yml file sets up the "environment" to run the tests
using the test containers. It sets up the network, starts the
containers, sets up the moint point, the port mappings, and the run
commands to build the new code.

For the postgresql, we use environment variables to setup the user,
password and database.

For the web container, it builds the new code and copies some css, js,
and fonts.

For the tomcat container, it builds the new code, pulls new dependencies
and deployes all that plus other needed files into the tomcat webapp
directory.

Signed-off-by: Jordi Massaguer Pla <jmassaguerpla@suse.de>