CI: registry visible from ubuntu runner
When we run the tests that use docker to build containers, docker is run inside
a container that runs with podman. The trick to make this work, is to
mount the docker socket in that container. This means, that the docker daemon
is run in the ubuntu runner, the "podman host".

Thus, for the docker daemon to access the container registries, we need to
expose the ports from the container registry containers, to localhost, so
the docker daemon can pull from them.

Signed-off-by: Jordi Massaguer Pla <jmassaguerpla@suse.com>